+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Add JSON helper to show stringify objects (#121)
 - Update to Grafana 9.3.1 (#122)
-- Update Documentation links (#130)
+- Update Documentation links (#130, #131)
 
 ## 2.1.0 (2022-11-27)