+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Features / Enhancements
 
-- Add JSON helper to show stringify objects (#121)
+- Add JSON helper to show objects and arrays (#121)
 - Update to Grafana 9.3.1 (#122)
 - Update Documentation links (#130, #131)
+- Add Split Helper (#132)
+- Refactor Text Component and Styles (#133)
 
 ## 2.1.0 (2022-11-27)