+++ CHANGELOG.md
@@ -9,8 +9,9 @@
 - Update Documentation links (#130, #131)
 - Add Split Helper (#132)
 - Refactor Text Component and Styles (#133)
-- Add Javascript Code to add Handlebars helpers and Event handlers (#134)
-- Update default Content to `{{json @root}}` and Code Editor height to `200px`.
+- Add JavaScript Code to add Handlebars helpers and Event handlers (#134)
+- Update default Content to `{{json @root}}` and Code Editor height to `200px` (#134)
+- Update CSS to fit images to screen (#135)
 
 ## 2.1.0 (2022-11-27)